Can work

+++ Player.cpp
@@ -24,7 +24,7 @@ vector<Move> Player::genAllMoves( Board & board)
 	return ret;
 }
 
-int Player::eval( Board& board){
+float Player::eval( Board& board){
     //Pieces::getDistanceToEnemyBase() 获取棋子离敌方base的距离
     //pieces::getChessPower() 获取子力
